+++ tala/model/input_hypothesis.py
@@ -1,14 +1,20 @@
+from typing import Text  # noqa: F401
+
+
 class InputHypothesis(object):
     def __init__(self, utterance, confidence):
+        # type: (Text, float) -> None
         self._utterance = utterance
         self._confidence = confidence
 
     @property
     def utterance(self):
+        # type: () -> Text
         return self._utterance
 
     @property
     def confidence(self):
+        # type: () -> float
         return self._confidence
 
     def __eq__(self, other):