+++ tala/model/user_move.py
@@ -1,23 +1,31 @@
+from typing import Text  # noqa: F401
+
+
 class UserMove(object):
     def __init__(self, semantic_expression, perception_confidence, understanding_confidence):
+        # type: (Text, float, float) -> None
         self._semantic_expression = semantic_expression
         self._perception_confidence = perception_confidence
         self._understanding_confidence = understanding_confidence
 
     @property
     def is_ddd_specific(self):
+        # type: () -> bool
         return False
 
     @property
     def semantic_expression(self):
+        # type: () -> Text
         return self._semantic_expression
 
     @property
     def perception_confidence(self):
+        # type: () -> float
         return self._perception_confidence
 
     @property
     def understanding_confidence(self):
+        # type: () -> float
         return self._understanding_confidence
 
     def __eq__(self, other):
@@ -42,15 +50,18 @@ class UserMove(object):
 
 class DDDSpecificUserMove(UserMove):
     def __init__(self, ddd, semantic_expression, perception_confidence, understanding_confidence):
+        # type: (Text, Text, float, float) -> None
         super(DDDSpecificUserMove, self).__init__(semantic_expression, perception_confidence, understanding_confidence)
         self._ddd = ddd
 
     @property
     def is_ddd_specific(self):
+        # type: () -> bool
         return True
 
     @property
     def ddd(self):
+        # type: () -> Text
         return self._ddd
 
     def __unicode__(self):